backend/requirements_hf.txt
gradio==4.44.0
tensorflow==2.15.0
pillow==12.1.1
numpy==1.26.4